+++ tests/Active/Fixture/SoftDeleteUser.php
@@ -5,8 +5,9 @@ namespace Rougin\Ezekiel\Active\Fixture;
 use Rougin\Ezekiel\Active\Model;
 
 /**
- * @property integer $id
- * @property string  $name
+ * @property integer     $id
+ * @property string      $name
+ * @property string|null $deleted_at
  *
  * @package Ezekiel
  *