+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to `Ezekiel` will be documented in this file.
 
+## 0.1.3 - 2026-06-26
+
+### Added
+- `whereNull`, `whereNotNull` in `Builder`, `Model`
+- `trashed`, `restore`, `forceDelete` in `Model`
+
+### Changed
+- Improved documentation
+
 ## 0.1.2 - 2026-06-21
 
 ### Fixed